updates cvrpsep files

+++ include/ext/cvrpsep/fcapfix.h
@@ -1,3 +1,4 @@
+/* SAS modified this file. */
 /* (C) Copyright 2003 Jens Lysgaard. All rights reserved. */
 /* OSI Certified Open Source Software */
 /* This software is licensed under the Common Public License Version 1.0 */
@@ -7,7 +8,7 @@
 
 void FCAPFIX_ComputeCuts(ReachPtr SupportPtr,
                          int NoOfCustomers,
-                         int *Demand, int CAP,
+                         double *Demand, double CAP,
                          int *SuperNodeSize,
                          double **XMatrix,
                          int MaxCuts,
