+++ include/ext/cvrpsep/memmod.h
@@ -1,3 +1,4 @@
+/* SAS modified this file. */
 /* (C) Copyright 2003 Jens Lysgaard. All rights reserved. */
 /* OSI Certified Open Source Software */
 /* This software is licensed under the Common Public License Version 1.0 */
@@ -6,9 +7,10 @@
 #define _H_MEMMOD
 
 #include <malloc.h>
+#include <inttypes.h>
 
-void* MemGet(int NoOfBytes);
-void* MemReGet(void *p, int NewNoOfBytes);
+void* MemGet(long unsigned int NoOfBytes); 
+void* MemReGet(void *p, long unsigned int NewNoOfBytes);
 void MemFree(void *p);
 char* MemGetCV(int n);
 char** MemGetCM(int Rows, int Cols);
